Write initial README and add a skeleton command
With this commit I have explained how to install the framework
and tried to illustrate why I have started this and what is
behind it in the sense of technology.

I hope that it is clear enough that this project is still in its
infancy and a lot of stuff does not function yet because I
am working on it.

This is part of issue https://github.com/somos/framework/issues/26 but does not completely fix it because a usage section is still missing